+++ src/index.ts
@@ -81,6 +81,7 @@ import Switch, { SwitchToggle } from './Molecules/Switch';
 import Tabs from './Molecules/Tabs';
 import TabsNav from './Molecules/TabsNav';
 import Timeline from './Molecules/Timeline';
+import Toggle from './Molecules/Toggle';
 import Tooltip from './Molecules/Tooltip';
 import TruncateWithTooltip from './Molecules/TruncateWithTooltip';
 import QuickFilter from './Molecules/QuickFilter';
@@ -174,13 +175,13 @@ export {
   Modal,
   MultiStepProgress,
   Number,
-  /** @deprecated */ OldIcon,
+  /** @deprecated use Icon instead */ OldIcon,
   PageHeaderCard,
   PageWrapper,
   Pagination,
   PersistentTooltip,
   Pill,
-  /** @deprecated */ PillButton,
+  /** @deprecated use Button.Pill instead */ PillButton,
   Portal,
   ProgressBar,
   ProgressIndicator,
@@ -195,7 +196,7 @@ export {
   Slider,
   Spinner,
   StatusModal,
-  Switch,
+  /** @deprecated use Toggle instead */ Switch,
   SwitchToggle,
   TabTitle,
   Table,
@@ -208,6 +209,7 @@ export {
   Thead,
   Time,
   Timeline,
+  Toggle,
   Tooltip,
   Tr,
   TrackingContext,